moves status magic to separate files
